remove useless part , rename some variables/functions

+++ src/nanokernel/inner/__nanokernel_task.h
@@ -6,9 +6,9 @@
 #include "../Misc/src/definitions.h"
 
 struct __nanokernel_Task_t {
-    int32_t* stack_ptr;
-    int32_t* stack_start;
-    int32_t* stack_end;
+    intptr_t* stack_ptr;
+    intptr_t* stack_start;
+    intptr_t* stack_end;
     size_t stack_size;
     int8_t id;
     long priority;
